refactor: requests arch
