+++ .env.production
@@ -1,5 +1,6 @@
 SOLANA_NETWORK=mainnet
-RPC_ENDPOINT=https://wild-red-morning.solana-mainnet.quiknode.pro/e48180a05f9f7ab63b6d9f0609f0ba675854e471/
+#RPC_ENDPOINT=https://wild-red-morning.solana-mainnet.quiknode.pro/e48180a05f9f7ab63b6d9f0609f0ba675854e471/
+RPC_ENDPOINT=https://polished-fragrant-dawn.solana-mainnet.quiknode.pro/8005e8943672dd7c0a751fe88526a6cca7954072/
 BACKEND_DOMAIN=api.hadeswap.com
 PROGRAM_PUBKEY=hadeK9DLv9eA7ya5KCTqSvSvRZeJC3JgD5a9Y3CNbvu
 PROTOCOL_FEE_PUBKEY=My11111111111111111111111111111111111111111
