+++ src/requests/constants.ts
@@ -1,0 +1,2 @@
+export const BASE_STALE_TIME = 5 * 60 * 1000; // 5 min
+export const SHORT_STALE_TIME = 10000; // 10 sec
